Remove Formula definition from init
