+++ src/energy_cost/formula/formula.py
@@ -1,44 +1,43 @@
-from __future__ import annotations
+from typing import Annotated, cast
 
-import datetime as dt
-from abc import ABC, abstractmethod
-from datetime import UTC
+from pydantic import Discriminator, Tag
 
-import pandas as pd
-from pydantic import BaseModel
-
-from energy_cost.resolution import Resolution, align_timestamps_to_tz, detect_resolution_and_range
+from .index import IndexFormula
+from .periodic import PeriodicFormula
+from .scheduled import ScheduledFormulas
+from .tiered import TieredFormula
 
 
-class FormulaBase(ABC, BaseModel):
-    @abstractmethod
-    def get_values(
-        self,
-        start: dt.datetime,
-        end: dt.datetime,
-        resolution: Resolution,
-        timezone: dt.tzinfo = UTC,
-    ) -> pd.DataFrame:
-        """Return time-indexed values for the formula."""
+def _formula_discriminator(v: object) -> str | None:
+    """Return the ``kind`` tag used for union routing.
 
-    def apply(
-        self,
-        data: pd.DataFrame,
-        resolution: Resolution | None = None,
-        timezone: dt.tzinfo = UTC,
-        *,
-        start: dt.datetime | None = None,
-        end: dt.datetime | None = None,
-    ) -> pd.DataFrame:
-        """Apply formula values to a dataframe of quantities and return a single value column."""
-        if data.empty:
-            return data.copy()
-        data = align_timestamps_to_tz(data, timezone)
-        if start is None or end is None or resolution is None:
-            start, end, resolution = detect_resolution_and_range(data, resolution)
-        formula_values = self.get_values(start, end, resolution, timezone)
+    Explicit ``kind`` fields take priority; otherwise heuristics based on the
+    presence of characteristic keys are used (for YAML that omits ``kind``).
+    Model instances are routed via their ``.kind`` attribute.
+    """
+    if isinstance(v, dict):
+        d = cast(dict[str, object], v)
+        if "kind" in d:
+            return str(d["kind"])
+        if "bands" in d:
+            return "tiered"
+        if "period" in d:
+            return "periodic"
+        if "schedule" in d:
+            return "scheduled"
+        if "constant_cost" in d or "variable_costs" in d:
+            return "index"
+        return None
+    return getattr(v, "kind", None)
 
-        result = data.reset_index(drop=True)
-        formula_series = formula_values.set_index("timestamp")["value"].reindex(result["timestamp"])
-        result["value"] = formula_series.values * result["value"].values
-        return result
+
+# Discriminated union of all concrete formula types.
+# A callable Discriminator handles both explicit ``kind`` fields and
+# heuristic routing for YAML that omits ``kind``.
+Formula = Annotated[
+    Annotated[IndexFormula, Tag("index")]
+    | Annotated[PeriodicFormula, Tag("periodic")]
+    | Annotated[ScheduledFormulas, Tag("scheduled")]
+    | Annotated[TieredFormula, Tag("tiered")],
+    Discriminator(_formula_discriminator),
+]
