+++ src/energy_cost/formula/scheduled.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel, Field, model_validator
 
 from energy_cost.resolution import Resolution, align_datetime_to_tz, to_pandas_freq
 
-from .formula import FormulaBase
+from .base import FormulaBase
 
 if TYPE_CHECKING:
     from energy_cost.formula import Formula
