+++ src/energy_cost/formula/tiered.py
@@ -16,7 +16,7 @@ from energy_cost.resolution import (
     to_pandas_freq,
 )
 
-from .formula import FormulaBase
+from .base import FormulaBase
 
 if TYPE_CHECKING:
     from energy_cost.formula import Formula
